// src/generated-api/src/models/DomainOrganizationWithExtraInfo.ts
export interface DomainOrganizationWithExtraInfo {
    /**
     * 
     * @type {DomainAdvertiserExtraInfo}
     * @memberof DomainOrganizationWithExtraInfo
     */
    advertiserExtraInfo?: DomainAdvertiserExtraInfo;
    /**
     * 
     * @type {DomainAffiliateExtraInfo}
     * @memberof DomainOrganizationWithExtraInfo
     */
    affiliateExtraInfo?: DomainAffiliateExtraInfo;
    /**
     * 
     * @type {string}
     * @memberof DomainOrganizationWithExtraInfo
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainOrganizationWithExtraInfo
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof DomainOrganizationWithExtraInfo
     */
    organizationId?: number;
    /**
     * 
     * @type {DomainOrganizationType}
     * @memberof DomainOrganizationWithExtraInfo
     */
    type?: DomainOrganizationType;
    /**
     * 
     * @type {string}
     * @memberof DomainOrganizationWithExtraInfo
     */
    updatedAt?: string;
}